+++ lib/whois/record/parser/whois.denic.de.rb
@@ -21,7 +21,7 @@ module Whois
       # @author Aaron Mueller <mail@aaron-mueller.de>
       #
       class WhoisDenicDe < Base
-        include Scanners::Ast
+        include Scanners::Nodable
 
         property_supported :disclaimer do
           node("Disclaimer")