+++ lib/whois/record/parser/whois.nic.cc.rb
@@ -17,7 +17,7 @@ module Whois
 
       # Parser for the whois.nic.cc server.
       class WhoisNicCc < Base
-        include Scanners::Ast
+        include Scanners::Nodable
 
         property_supported :disclaimer do
           node("Disclaimer")