+++ lib/whois/record/parser/whois.nic.dm.rb
@@ -22,7 +22,7 @@ module Whois
       #
       # @since  RELEASE
       class WhoisNicDm < Base
-        include Scanners::Ast
+        include Scanners::Nodable
 
         property_supported :disclaimer do
           node("field:disclaimer")