+++ lib/whois/record/parser/whois.nic.tv.rb
@@ -17,7 +17,7 @@ module Whois
 
       # Parser for the whois.nic.tv server.
       class WhoisNicTv < Base
-        include Scanners::Ast
+        include Scanners::Nodable
 
         property_supported :disclaimer do
           node("Disclaimer")